Remove lots of old cruft
- We've removed almost all of the common functionality that is no longer
  used. Pretty much 80% of nodeobs_common.cpp has been wiped out (which
  was ~4000 lines literally halving our binary size).

- Removed the old index and obspp manager as they're no longer used.

- Some WIP in moving things that belong in implementation module out
  of the corresponding headers (specifically in autoconfig and service).
