+++ CMakeLists.txt
@@ -75,7 +75,6 @@ set(PROJECT_SOURCE
 	"src/obspp/obspp-display.cpp" "src/obspp/obspp-display.hpp"
 	"src/obspp/obspp-encoder.cpp" "src/obspp/obspp-encoder.hpp"
 	"src/obspp/obspp-filter.cpp" "src/obspp/obspp-filter.hpp"
-	"src/obspp/obspp-index.hpp"
 	"src/obspp/obspp-input.cpp" "src/obspp/obspp-input.hpp"
 	"src/obspp/obspp-output.cpp" "src/obspp/obspp-output.hpp"
 	"src/obspp/obspp-module.cpp" "src/obspp/obspp-module.hpp"
@@ -91,13 +90,10 @@ set(PROJECT_SOURCE
 	"src/nodeobs_api.h" "src/nodeobs_api.cpp"
 	"src/nodeobs_audio_encoders.h" "src/nodeobs_audio_encoders.cpp"
 	"src/nodeobs_autoconfig.h" "src/nodeobs_autoconfig.cpp"
-	"src/nodeobs_common.h" "src/nodeobs_common.cpp"
+	"src/nodeobs_common.cpp"
 	"src/nodeobs_display.h" "src/nodeobs_display.cpp"
-	"src/nodeobs_event.h" "src/nodeobs_event.cpp"
 	"src/nodeobs_service.h" "src/nodeobs_service.cpp"
 	"src/nodeobs_settings.h" "src/nodeobs_settings.cpp"
-	"src/nodeobs_obspp_manager.hpp" "src/nodeobs_obspp_manager.cpp"
-	"src/nodeobs_obspp_index.hpp" "src/nodeobs_obspp_index.cpp"
 	"src/nodeobs_content.h"
 	"src/gs-limits.h"
 	"src/gs-vertex.h" "src/gs-vertex.cpp"
