+++ src/nodeobs_async.h
@@ -8,8 +8,6 @@
 #include <uv.h>
 #include <v8.h>
 
-#include "nodeobs_obspp_manager.hpp"
-
 namespace obs {
 
 class ForeignWorker {
@@ -93,54 +91,5 @@ struct CallbackQueue {
         uv_async_send(&async);
     }
 };
-
-static void check_type(obs::object &object, obs::object::objectType type)
-{
-	if (!object.testType(type)) {
-		std::string error("object::check_type failed ");
-		error += object_type_map[object.type];
-		error += " and ";
-		error += object_type_map[type];
-		Nan::ThrowTypeError(error.c_str());
-	}
-}
-
-static void set_id(v8::Isolate *isolate, v8::Local<v8::Object> object, uint32_t id)
-{
-    v8::HandleScope scope(isolate);
-    object->Set(
-        v8::String::NewFromUtf8(isolate, "id"),
-        v8::Uint32::New(isolate, id));
-}
-
-static uint32_t get_id(v8::Isolate *isolate, v8::Local<v8::Object> object)
-{
-    v8::HandleScope scope(isolate);
-    v8::Local<v8::Value> value =
-        object->Get(v8::String::NewFromUtf8(isolate, "id"));
-
-    return value->ToUint32()->Value();
-}
-
-template <obs::object::objectType Type, typename Handle>
-Handle get_handle(uint32_t id)
-{
-	if (g_objectManager.isUsed(id, false)) {
-		return nullptr;
-	}
-
-    obs::object object = g_objectManager.getObject(id);
-
-    check_type(object, Type);
-    return reinterpret_cast<Handle>(object.handle);
-}
-
-template <obs::object::objectType Type, typename Handle>
-Handle get_handle(v8::Isolate *isolate, v8::Local<v8::Object> object)
-{
-    uint32_t id = get_id(isolate, object);
-
-    return get_handle<Type, Handle>(id);
-}
 
 }
