+++ src/nodeobs_service.cpp
@@ -87,17 +87,7 @@ void OBS_service::OBS_service_startRecording(const FunctionCallbackInfo<Value>& 
 
 void OBS_service::OBS_service_stopStreaming(const FunctionCallbackInfo<Value>& args)
 {
-	Isolate *isolate = args.GetIsolate();
-
-	if (args[0]->IsUndefined() && !args[0]->IsBoolean()) {
-		isolate->ThrowException(
-			v8::Exception::SyntaxError(
-				v8::String::NewFromUtf8(isolate, "{forceStop} is not a <boolean>!")
-			)
-		);
-		return;
-	}
-	stopStreaming(args[0]->BooleanValue());
+	stopStreaming();
 }
 
 void OBS_service::OBS_service_stopRecording(const FunctionCallbackInfo<Value>& args)
@@ -142,513 +132,6 @@ void OBS_service::OBS_service_isStreamingOutputActive(const FunctionCallbackInfo
     int isActive = isStreamingOutputActive();
 
     args.GetReturnValue().Set(String::NewFromUtf8(isolate, to_string(isActive).c_str()));
-}
-
-void OBS_service::OBS_service_test_resetAudioContext(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-    if(resetAudioContext()) {
-        result = "SUCCESS";
-    } else {
-        result = "FAILURE";
-    }
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_resetVideoContext(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-    if(resetVideoContext(NULL) == OBS_VIDEO_SUCCESS) {
-        result = "SUCCESS";
-    } else {
-        result = "FAILURE";
-    }
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_createAudioEncoder(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-
-    createAudioEncoder();
-
-    if(audioEncoder != NULL) {
-        result = "SUCCESS";
-    } else {
-        result = "FAILURE";
-    }
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_createVideoStreamingEncoder(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-
-    createVideoStreamingEncoder();
-
-    if(videoStreamingEncoder != NULL) {
-        result = "SUCCESS";
-    } else {
-        result = "FAILURE";
-    }
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_createVideoRecordingEncoder(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-
-    createVideoRecordingEncoder();
-
-    if(videoRecordingEncoder != NULL) {
-        result = "SUCCESS";
-    } else {
-        result = "FAILURE";
-    }
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_createService(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-    createService();
-    if(service != NULL) {
-        result = "SUCCESS";
-    } else {
-        result = "FAILURE";
-    }
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_createRecordingSettings(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-    if(createRecordingSettings() != NULL) {
-        result = "SUCCESS";
-    } else {
-        result = "FAILURE";
-    }
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_createStreamingOutput(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-    
-    createStreamingOutput();
-
-    if(streamingOutput != NULL) {
-        result = "SUCCESS";
-    } else {
-        result = "FAILURE";
-    }
-
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_createRecordingOutput(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-
-    createRecordingOutput();
-
-    if(recordingOutput != NULL) {
-        result = "SUCCESS";
-    } else {
-        result = "FAILURE";
-    }
-
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_startStreaming(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-
-    //Create output
-    createStreamingOutput();
-    
-    //Creating video encoder
-    createVideoStreamingEncoder();
-    
-    //Creating audio encoder
-    createAudioEncoder();
-    
-    //Reset audio
-    resetAudioContext();
-    
-    //Reset video
-    resetVideoContext(NULL);
-    
-    //Associate A/V encoders with current context
-    associateAudioAndVideoToTheCurrentStreamingContext();
-    
-    //Load service
-    createService();
-    
-    //Update stream output settings
-    associateAudioAndVideoEncodersToTheCurrentStreamingOutput();
-    
-    //Assign service to an ouput
-    setServiceToTheStreamingOutput();
-
-
-    string result;
-    if(startStreaming()) {
-        result = "SUCCESS";
-    } else {
-        result = "FAILURE";
-    }
-
-    int timeout = 8;
-
-    while(!obs_output_active(streamingOutput)) {
-		std::this_thread::sleep_for(std::chrono::milliseconds(500));
-		//usleep(500000);
-		timeout--;
-		if(timeout == 0){
-			result = "FAILURE";
-			args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-			return;
-		}
-    }
-
-    stopStreaming(false);
-
-    timeout = 8;
-    while(obs_output_active(streamingOutput)) {
-		std::this_thread::sleep_for(std::chrono::milliseconds(500));
-		//usleep(500000);
-		timeout--;
-		if(timeout == 0){
-			args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-			return;
-		}
-    }
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_startRecording(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-
-    createRecordingOutput();
-    
-    //Creating video encoder
-    createVideoRecordingEncoder();
-
-	//Creating audio encoder
-	createAudioEncoder();
-    
-    //Reset audio
-    resetAudioContext();
-    
-    //Reset video
-    resetVideoContext(NULL);
-    
-    //Associate A/V encoders with current context
-    associateAudioAndVideoToTheCurrentRecordingContext();
-    
-    //Update stream output A/V encoders
-    associateAudioAndVideoEncodersToTheCurrentRecordingOutput();
-
-    //Setting recording output settings
-    setRecordingSettings();
-
-    string result;
-    if(startRecording()) {
-        result = "SUCCESS";
-    } else {
-        result = "FAILURE";
-    }
-
-    int timeout = 8;
-
-    while(!obs_output_active(recordingOutput)) {
-		std::this_thread::sleep_for(std::chrono::milliseconds(500));
-		//usleep(500000);
-		timeout--;
-		if(timeout == 0){
-			result = "FAILURE";
-			args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-			return;
-		}
-    }
-
-    stopRecording();
-
-    timeout = 8;
-    while(obs_output_active(recordingOutput)) {
-		std::this_thread::sleep_for(std::chrono::milliseconds(500));
-		//usleep(500000);
-		timeout--;
-		if(timeout == 0){
-			args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-			return;
-		}
-    }
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_stopStreaming(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-
-	//Create output
-    createStreamingOutput();
-    
-    //Creating video encoder
-    createVideoStreamingEncoder();
-    
-    //Creating audio encoder
-    createAudioEncoder();
-    
-    //Reset audio
-    resetAudioContext();
-    
-    //Reset video
-    resetVideoContext(NULL);
-    
-    //Associate A/V encoders with current context
-    associateAudioAndVideoToTheCurrentStreamingContext();
-    
-    //Load service
-    createService();
-    
-    //Update stream output settings
-    associateAudioAndVideoEncodersToTheCurrentStreamingOutput();
-    
-    //Assign service to an ouput
-    setServiceToTheStreamingOutput();
-
-    startStreaming();
-
-	int timeout = 8;
-
-    while(!obs_output_active(streamingOutput)) {
-		std::this_thread::sleep_for(std::chrono::milliseconds(500));
-		//usleep(500000);
-		timeout--;
-		if(timeout == 0){
-			result = "FAILURE";
-			args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-			return;
-		}
-    }
-
-    stopStreaming(false);
-
-    timeout = 8;
-    while(obs_output_active(streamingOutput)) {
-		std::this_thread::sleep_for(std::chrono::milliseconds(500));
-		//usleep(500000);
-		timeout--;
-		if(timeout == 0){
-			result = "FAILURE";
-			args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-			return;
-		}
-    }
-
-    if(!obs_output_active(streamingOutput)) {
-        result = "SUCCESS";
-    } else {
-        result = "FAILURE";
-    }
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_stopRecording(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-
-    createRecordingOutput();
-    
-    //Creating video encoder
-    createVideoRecordingEncoder();
-
-	//Creating audio encoder
-	createAudioEncoder();
-    
-    //Reset audio
-    resetAudioContext();
-    
-    //Reset video
-    resetVideoContext(NULL);
-    
-    //Associate A/V encoders with current context
-    associateAudioAndVideoToTheCurrentRecordingContext();
-    
-    //Update stream output A/V encoders
-    associateAudioAndVideoEncodersToTheCurrentRecordingOutput();
-
-    //Setting recording output settings
-    setRecordingSettings();
-
-
-    startRecording();
-
-	int timeout = 8;
-
-    while(!obs_output_active(recordingOutput)) {
-		std::this_thread::sleep_for(std::chrono::milliseconds(500));
-		//usleep(500000);
-		timeout--;
-		if(timeout == 0){
-			result = "FAILURE";
-			args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-			return;
-		}
-    }
-
-    stopRecording();
-
-    timeout = 8;
-    while(obs_output_active(recordingOutput)) {
-		std::this_thread::sleep_for(std::chrono::milliseconds(500));
-		//usleep(500000);
-		timeout--;
-		if(timeout == 0){
-			result = "FAILURE";
-			args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-			return;
-		}
-    }
-
-    if(!obs_output_active(recordingOutput)) {
-        result = "SUCCESS";
-    } else {
-        result = "FAILURE";
-    }
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_associateAudioAndVideoToTheCurrentStreamingContext(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-
-    associateAudioAndVideoToTheCurrentStreamingContext();
-
-    video_t* video = obs_encoder_video(videoStreamingEncoder);
-    audio_t* audio = obs_encoder_audio(audioEncoder);
-
-    if(video == obs_get_video() && audio == obs_get_audio()) {
-        result = "SUCCESS";
-    }
-    else {
-        result = "FAILURE";
-    }
-
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_associateAudioAndVideoToTheCurrentRecordingContext(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-
-    associateAudioAndVideoToTheCurrentRecordingContext();
-
-    video_t* video = obs_encoder_video(videoRecordingEncoder);
-    audio_t* audio = obs_encoder_audio(audioEncoder);
-
-    if(video == obs_get_video() && audio == obs_get_audio()) {
-        result = "SUCCESS";
-    }
-    else {
-        result = "FAILURE";
-    }
-
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_associateAudioAndVideoEncodersToTheCurrentStreamingOutput(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-
-    associateAudioAndVideoEncodersToTheCurrentStreamingOutput();
-
-    obs_encoder_t* currentVideoEncoder = obs_output_get_video_encoder(streamingOutput);
-    obs_encoder_t* currentAudioEncoder = obs_output_get_audio_encoder(streamingOutput, 0);
-
-    if(currentVideoEncoder == videoStreamingEncoder && currentAudioEncoder == audioEncoder) {
-        result = "SUCCESS";
-    }
-    else {
-        result = "FAILURE";
-    }
-
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_associateAudioAndVideoEncodersToTheCurrentRecordingOutput(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-
-    associateAudioAndVideoEncodersToTheCurrentRecordingOutput();
-
-    obs_encoder_t* currentVideoEncoder = obs_output_get_video_encoder(recordingOutput);
-    obs_encoder_t* currentAudioEncoder = obs_output_get_audio_encoder(recordingOutput, 0);
-
-    if(currentVideoEncoder == videoRecordingEncoder && currentAudioEncoder == audioEncoder) {
-        result = "SUCCESS";
-    }
-    else {
-        result = "FAILURE";
-    }
-
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_setServiceToTheStreamingOutput(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-
-    setServiceToTheStreamingOutput();
-    obs_service_t* currentService = obs_output_get_service(streamingOutput);
-
-    if(currentService == service) {
-        result = "SUCCESS";
-    }
-    else {
-        result = "FAILURE";   
-    }
-
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
-}
-
-void OBS_service::OBS_service_test_setRecordingSettings(const FunctionCallbackInfo<Value>& args)
-{
-    Isolate* isolate = args.GetIsolate();
-    string result;
-
-    setRecordingSettings();
-    
-    obs_data_t* currentSettings = obs_output_get_settings(recordingOutput);
-
-    if(currentSettings != NULL) {
-        result = "SUCCESS";
-    }
-    else {
-        result = "FAILURE";
-    }
-    
-    args.GetReturnValue().Set(String::NewFromUtf8(isolate, result.c_str()));
 }
 
 void LoadAudioDevice(const char *name, int channel, obs_data_t *parent)
@@ -1307,12 +790,9 @@ bool OBS_service::startRecording(void)
 	return obs_output_start(recordingOutput);
 }
 
-void OBS_service::stopStreaming(bool forceStop)
+void OBS_service::stopStreaming(void)
 {
-	if(forceStop)
-		obs_output_force_stop(streamingOutput);
-	else
-		obs_output_stop(streamingOutput);
+	obs_output_stop(streamingOutput);
 }
 
 void OBS_service::stopRecording(void)
@@ -2148,13 +1628,7 @@ void OBS_service::updateStreamSettings(void)
 		}
     }
 
-	bool useDelay = config_get_bool(config, "Output", "DelayEnable");
-	int delaySec = config_get_int(config, "Output", "DelaySec");
-	bool preserveDelay = config_get_bool(config, "Output", "DelayPreserve");
-
-	obs_output_set_delay(streamingOutput, useDelay ? delaySec : 0,
-		preserveDelay ? OBS_OUTPUT_DELAY_PRESERVE : 0);
-
+    resetVideoContext("Stream");
     associateAudioAndVideoToTheCurrentStreamingContext();
     associateAudioAndVideoEncodersToTheCurrentStreamingOutput();
 }
@@ -2180,6 +1654,7 @@ void OBS_service::updateRecordSettings(void)
         updateAdvancedRecordingOutput();
     }
 
+    resetVideoContext("Record");
     associateAudioAndVideoToTheCurrentRecordingContext();
     associateAudioAndVideoEncodersToTheCurrentRecordingOutput();
 }
